backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.3
numpy==1.26.0
requests==2.31.0
pandas==2.1.0
sentence-transformers==2.2.2
flask-cors==4.0.0
openai==1.12.0 